[DF] Custom columns managed locally by nodes
Before this commit only the root node, the RLoopManager, kept the list of the custom columns defined by the user. This meant that it was not possible to define two columns with the same names in two different branches.

After this commit, each node in the graph has a map with pointers to all the columns defined up to that node. Nodes in different branches may see different custom column. This allows operations like the following to work:

 ROOT::RDataFrame d(3);
   auto branch1 = d.Define("b2", []() { return 1; });
   auto branch2 = d.Define("b2", []() { return 2; });

- Improved overall code quality
- Fixed possible missing initialization of the custom columns. Now the loop manager has again an observer pointer to all the custom columns, only to initialize them.